added three swift helpers
new ios / osx class + DEMO:
Makes observing a notificationCenter easy, convenient and more swift like.

new osx entension:
Category for NSWindow that provides helpers to fade in/out a window when ordering it in/out

new osx entension:
a Category on NSImage to get NSData in PNG Format for saving it to file. This simplifies the 'weirdness' of NSImageRepresentations away ;)

Signed-off-by: Dominik Pich <Dominik@pich.info>